Standardize Yaesu checksum calculation
This patch introduces a YaesuChecksum object that represents a checksum
to be calculated with a range and address to store the result.  Radio drivers
inheriting from YaesuCloneModeRadio should now implement the _checksums()
method and return a list of YaesuChecksum objects describing where and how
to calculate their values.  The main yaesu_clone code now uses these
objects to calculate the checksums to (a) update them before uploading,
and (b) check them after downloading.