Add an automated test harness
This is in tests/run_tests.  It is a (hopefully) modular approach
to banging on the radio drivers in a generic way to find/report
issues.  It attempts to gracefully handle failues and crashes,
logging information to logs/ when available and appropriate.

Two test cases are included:

BruteForce, which attempts to try to set every possible value
for various things like tone, duplex, skip, etc.  It should be
expanded, but covers some of the juicy stuff right now.

CopyAll, which attempts to do an effective import operation from
a CSV file (images/csv.csv) which has a bunch of gnarly stuff in
it (such as 12.5kHz step channels, D-STAR, 800MHz, etc).

Each of these are run against every (clone mode) driver in the
directory, making sure that any memories that we set come back
as we expect them.