Allow querying network sources directly (Feature #126)
This change puts the "Import from..." network source menu items
inside a generic "Import from data source" menu item and adds
another called "Query data source". It is populated with similar
items for each of the network data sources currently supported.
When the user selects one of these, the usual query process happens,
but in the end, we open the result in a read-only tab so that they
can browse, copy, and paste selections into another tab.