[uv5r] Read the auxiliary block before uploading to verify firmware compat
Thanks to investigation by Jim Unroe, it looks like the version stored
in the aux block is non-volatile and reliable for version detection.
This change checks that block against that of the image to make sure
that there is an exact match and refuses to upload if not. This could
potentially be made more tolerant by allowing the upload, but skipping
the aux block for non-firmware matching radios.

Addresses #344