[uv5r] Make the aux block upload conditional on firmware version
This loosens the check for a matching firmware version on upload.
Instead of requiring it to be the same for any upload, this makes
it not do the aux block upload if the versions don't match. Now,
users can still share images between radios, but must do a fresh
clone to get the aux block settings uploaded properly. An error
is raised after a successful main block upload, but before the
aux block to let them know that this happened.

Improves on #344