[icq7] Add support for "Auto" mode
This fixes #390 by adding support for auto modes in the ICQ7 driver. It
also adds support to the chirp core, and teaches the import logic how to
make (very rough) guesses about an appropriate mode when importing from a
radio that supports automatic mode selection to one that does not. It has
the side-effect of allowing Auto modes in CSV files.