[ftx800] Make the bank model smarter
This introduces a caching mechanism to the BankModel(s) used for the
FT7800 and FT8800 radios. It's probably something that will need to be
abstracted and used for any other NxN (i.e. Yaesu) radios, but for now
I'll just put it here.

It changes the time it takes to load the banks on the FT8800 from about
23 seconds on my machine to 0.43 seconds.

Note: Fix incorrect upper bound on the FT8800 incidental to the rest.

Related to #380