Fix improper mem.extra setting names
This is confusing but the "name" of a setting is actuall the identifier,
and the "shortname" is actually the descriptive human-readable name.

Bug #513 was filed because it appeared that the memory detail editor
had this reversed, but in fact, it was just reversed in some of the
implementations. This corrects the implementations.