Add AnyTone 5888UV support
Note that the upload/download process takes much longer than the OEM
software at this point. That is because the OEM software is optimizing
by checking to see what memories are in use and only down/uploading the
parts of the memory that it has to. This may be added in chirp at a
later date, but it's tricky to get correct.

Addresses #365