Add tools/checkpatch.sh for style compliance checking
This adds tools/checkpatch.sh which checks a given HG revision
for style compliance. We should call this as part of our tests,
have Jenkins call it, and keep it full of things we'd like to
enforce. Developers should run this on their patches (when
possible) before submitting.

Had this idea after the patch for #1277