[developer] Control fontsize of file browser display - #1697
Add user specification of the font size to use for the developer tools
file browser display, without affecting any other Chirp displays.
This is done with a new chirp.config integer property browser_fontsize
in the [developer] section. The default size is unchanged at 10.
Values between 4 and 144 are accepted; others result in the default.

Add documentation of browser_fontsize to the README.developers file.

#1697