[BUG] Fix Range Exclusion Issue with Bank Editor
Memory Bounds ranges are specified as inclusive of the first as well
as the last; while python 'range' is inclusive of the first and exclusive
of the last.

As a result the BankEditor was not including the final memory slot within
the devices capable bounds, this fixes the issue and aligns with the
behaviour of the MemoryEditor.

Bug #2175