Updated radio settings interface.
Due to the nested nature of the settings editor tree
view, labels in the tree generally become obscured
because the pane is too small.

This change addresses the issue by a) adding a
horizontal paned window which allows the tree view
to be manually resized while b) flattening the radio
featrure radio settings group to return a single list
of all top level radio settings groups.

Bug #2285