Fixing issues with memory property editor.
This patch attempts to address a couple issues with the layout of the
property page. Mainly the issues relating to the layout of the table
containing the widgets was skewed (on windows at least) where some entries
appeared much big than others.  The other issue was that the multiple property
editor was showing blank entries for no apparent reason.

It addresses an issue with the multiple memory editor where the
selector of elements which were intnded to be hidden was still being displayed.

It adds basic tooltip support to all widgets and not just the extra settings.

It also groups elements into two groups (tabs) "general" and "other"; the
former signifies the basic radio settings which appear in the memory editor
while the later is the set of "extra" radio settings.

Bug #2251