Attempting to speed up memory editor load times
It appears that the memory editor is unecessarily attempting to simulatanesuoly
prefetch all memories from the radio twice at startup.

This is causing a somewhat signifigant performance hit as the tree view is
performing a linear search of its store attempting to prevent
duplicate entries (see while loop in set_memory)

As a result the duplicate call to prefill was having no effect other than
forcing a lookup of the treestore and did not appear to have any other
bearing on the application.

Another seperate, but related, improvement would be to increase the overall
performance of the linear search of the treestore.

Bug #2369