Allow cpep8.py script to make exceptions (#2355)
This patch enhances the cpep8.py script with the ability to ignore a
list of style rules on per-file basis.  There are cases where it may be
too disruptive to eliminate all style issues, so this mechanism allows
the style checker to enforce the remainder of the style rules.  The new
--strict option will cause the script to ignore the exception list.

As a bonus, this change improves interactivity, so --verbose will cause
the name of the file to be printed as it is checked (rather than
printing one long list before it started).  The downside: the pep8
module catches KeyboardInterrpt, so it's hard impossible to stop once it
gets started.