Improve cpep8 scripts (#2355)
This patch adds prints to cpep8.py's scanning function.  It's useful to
know what files the script is scanning during an update, as the in-tree
virtualenv (or other files) may be included by accident.

This patch also fixes cpep8.sh to use the correct path to cpep8.py,
allowing it to be run from anywhere in the source tree.