[UV-5R] Support BF-A58 as Radio With 3 Power Levels
The initial support for the BF-A58 added it as a radio with only 2 power
levels. It has 3 power levels. This patch moves it to the group radios
having 3 power levels.

Note: This patch also requires removing "BFT" from the BASETYPE_UV5R
list and adding it to the BASETYPE_F8HP list. This was accidentally done
already in a patch that added UV-82HP support.

Related to  #2261