[PATCH][TK-760G] Fix the driver in windows and CH with no TX, fixes #2999
The driver was not working on windows, I take a while to learn by the
hard way that the timing on windows is different from linux & mac.

Now it works, fixed also the offset = off when you set a channel.

I have extended the edges of the bands when they are very close to
the ham bands, that's the way we use this radios in Cuba, for example:

The radios with the default sub-band segment 148-174 Mhz can be pulled
down to 144 Mhz in the lower edge with no problem, this mod works also
in UHF.