[IC-T8A] Fix frequency encoding and decoding. Fixes #577
The IC-T8 stores frequencies in three bcd bytes. The last nibble of these three
bytes would naturally be the least significant digit in KHz, but this leaves no
way to encode frequencies divisible by 2.5 KHz. To account for this, the IC-T8
deviates from its standard bcd encoding for the final nibble. Instead of the
nibble representing 1 KHz, it represents 500 Hz: 0x5 is 2.5 KHz, 0xA is 5 KHz,
and so on. This patch implements this encoding and decoding in Chirp.