[UI] Default "Hide Unused Fields" to enabled. Completes #1779
The orignal patch to add this feature missed one place the hide_unused value
is read from. This caused new installs to show the Hide Unused Fields box as
checked, but not actually enable the feature. This patch adds defaut=True to
that last remaining line of code so that this feature is enabled on the first
run of Chirp.