[PATCH][Kenwood Serie 60G] Widen the band limits for some radios, fixes #3813
The OEM software for this radios can program it out of it's default band limits, some of the models has ranges of operations close but outside the ham bands.

This patch widen the default band limits for this radios, allowing the programming of these for hamm radio use, some of the new limits are tested by me and others by Stephen GN5SWP, other are just guesses.