[wouxun] Prevent KGUVD1PRadio driver from over-matching on leixen 898S images (copy from same fix on KG816, in preparation for leixen 898S driver and image), fixes #3947
The KGUVD1 match isn't very discriminate; basically any 8k image with
lots of 0xff toward the end could match. This copies the filter used
in the KG816 driver. Specifically the driver was matching to the 898S
test image that will be added shortly, and this fixes it, but other
users could run into the problem as well.