[UV-5R] UV-82HP (An error has occurred, Unexpected response from radio)
The "ident" returned by the radios supported by the uv5r.py driver used to be
static. The "ident" of the UV-82HP radio (and possibly other recent Baofeng
radios) now changes based on the band limit settings. When setting a band limit
to "1" or and value ending in "01" (101, 201, 301, etc.), the value ends up in
the "ident" and is stripped by the code that was added to support the UV-6 with
the longer 12 byte "ident".

This patch uses a different strategy to support the UV-6 with longer "ident"
restoring the use of any selectable band limit value.

Bug #3987