Alinco DJ-G7: Try slower baud rate if default rate fails (#4355)
Alinco support is suggesting to use lower baud rates if the programming
software fails to communicate with the radio reliably. This patch
alters the code for the Alinco DJ-G7 radio such that if falls back to
19200 baud if the default of 57600 baud does not work. The user must
set his radio to 19200 baud for this mechanism to kick in. Dialog boxes
were added accordingly using the chirp_common.RadioPrompts() feature.