Fix long-standing test runner output bug
The unified logging code arrests stdout for the purposes of non-interactive
console logging to a file. This has been preventing proper console output
when running in the build system, making it harder to debug failures. This
adds a CHIRP_TESTENV override environment variable, and always sets it from
the test runner since we should always have interactive output from it.

Related to #5237