Fix Yaesu FT-8900 bank model disable
This reverts/fixes a previous patch that tried to delete the get_bank_model()
method from a class to avoid loading the banks tab for the 8900, which does
not support them. This fails the static analysis checker at least, and isn't
really the right way to do it. Just returning None will cause the compat routine
in chirp_common to do the right thing, so this patch changes it to just do
that.

Re-fix for #6293