[frs/gmrs] 2017 rule changes Fixes #5753
This adds the additional FRS channels (15-22) and expands GMRS channels include
all frequencies add in the 2017 rule changes.

It also adds the GMRS repeater channels with the prescribed +5.0MHz offset
(though you'll have to find the correct tones for your local GMRS repeaters).

After some feedback on the mailing list, I've chosen to number the channels
according to the FRS numbering, rather than try to preserve the obsolete
numbering, and I've chosen to name the GMRS repeater channels with both the
"550" style numbering that's common in repeater naming and the "15R" name that
makes it clear that it's receiving on channel 15.