Fix crash in anytone_iii when memobj isn't loaded
The new anytone_iii driver inspected the memory object during get_features(),
just to count the memories, which shouldn't be necessary. This papers over
that failure with the statically-defined value, but ideally, we'd just replace
the dynamic check with a static number. However, this fixes the build, and
is the least amount of change.

Related to #3941