Remove errant ui import from anytone_iii driver
The patch for #3941 imports a module from the UI for some reason, but
does not use it. This seems to have been breaking the PPA build because
we don't have gtk as a build-dep (for good reason). Drivers should be
independent of the UI and are not allowed to import things from it. Since
this driver wasn't actually using the import, removing it should be simple.

Fixes #8493