Add a brute force test for model detection
The regular per-image model detection tests are not very well-suited
for sussing out issues with one image being detected as multiple
models, and it can be fooled if the right one is returned first from
the directory. This adds a new unit test that runs through all the
images and tries to match_model() every driver in the directory against
them, asserting that not more than one matches any given file.

Also included is a fix to make the unittest adapter use helpfully-
named test images and makes the existing Detect test report the
filename of the image it is trying, for easier debug.

#0