[RT22S] Prepare for adding Retevis RT22S
The RT22S requires support that is nearly identical to that of the Radtel T18.
This patch makes changes to the T18 driver in order to make it ready to support
the addition of the Retevis RT22S.

No features or settings have been modified, added or removed by this patch.

related to #7179