[KT-8R] Settings tab empty for QYT KT-8R
The Settings tabs in CHIRP are empty when the KT-8R has TMR set to M+ABCD.
This driver was derived from the QYT mobile radios which only have 15 TMR
selections. This radio has a 16 TMR selections similar to the BTech mobile
radios so so setting it to the 16th choice puts it out-of-range based on the
current programming.

This patch adds 16 TMR choices support for the QyT KT-8R.

Fixes #9051