[UV-5X3] Add saved VFOs to BTECH UV-5X3
It was discovered that the BTECH UV-5X3 has 3 memories for each VFO display
line that can save frequencies, one for each of the 3 bands (VHF/VHF2/UHF).
Tapping the [BAND] button of the radio will cycle through the 3 memories
associated with the selected display line and place the currently selected
memory into the currently selected VFO.

This patch adds support to allow the viewing/editing of these 6 memories.

related to #9515