[UV-5X3] Fix for Unable to See Settings on BTECH UV-5X3
The BTECH UV-5X3 as shipped from the factory does not always have all 6 of the
sub-VFOs populated with frequencies. This was generating and error conditions
that was preventing the Settings tabs from being displayed.

This patch detects an unpopulated sub-VFO setting and populates the setting
with the appropriate low band limit frequency.

fixes #9617